fix:seleccion ruleta, fix:cruces se ejecutaba siempre, xmin double, gui mas compacta, CromosomaReal, CruceAritmetico, SeleccionEstocastico

+++ facil/CromosomaNDF4.java
@@ -2,7 +2,7 @@ package facil;
 
 public class CromosomaNDF4 extends CromosomaND {
 	double xmax;
-	int xmin;
+	double xmin;
 	private int numgenes;
 
 	public CromosomaNDF4(int ngenes) {
